"TMP filesystem ({}) is read-only!" Fix
TMP filesystem ({}) is read-only! was showing up when running on windows environment. May be wrong in removing the .resolve part, but found that Java was trying to locate a file that wasn't being created when the server/client was ran, resulting in the check failing. Removing the resolve still seems to pickup if a directory is read only so should be good still (hopefully lol)